requirements.txt
Flask==1.1.1
slackclient==2.5.0
slackeventsapi==2.1.0
Flask-SQLAlchemy==2.4.1
Flask-Migrate==2.5.2
psycopg2==2.8.4
gunicorn
pg8000
